problema-02/requirements.txt
requests==2.32.3
psycopg2-binary==2.9.12
python-dotenv==1.1.0